Now it should look more like a package
