+++ .github/workflows/node.yml
@@ -1,24 +1,24 @@
-name: Node.js CI
+name: Node.js
 
 on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+ push:
+  branches: [main]
+ pull_request:
+  branches: [main]
 
 jobs:
-  build:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        node-version: [10.x, 12.x, 14.x, 15.x]
-    steps:
-    - uses: actions/checkout@v2
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v2
-      with:
-        node-version: ${{ matrix.node-version }}
-    - run: npm i
-    - run: npm ci
-    - run: npm run build --if-present
-    - run: npm test
+ build:
+  runs-on: ubuntu-latest
+  strategy:
+   matrix:
+    node-version: [10.x, 12.x, 14.x, 15.x]
+  steps:
+   - uses: actions/checkout@v2
+   - name: Use Node.js ${{ matrix.node-version }}
+     uses: actions/setup-node@v2
+     with:
+      node-version: ${{ matrix.node-version }}
+   - run: npm i
+   - run: npm ci
+   - run: npm run build --if-present
+   - run: npm test
