+++ README.md
@@ -1,23 +1,26 @@
-# Official Api Module of Dumb bot list
-npm install dumb-botlistjs
-```
-const dbl1 = require("dumb-botlistjs");
-const dbl = new dbl1("TOKEN-HERE", client);
+# Dumb Bot API
+
+To instal this package run `npm install dumb-botlistjs`
+Current version: 1.0.2
+
+#### Example of usage
+```js
+const botlist = require("dumb-botlistjs");
+const dbl = new botlist("TOKEN-HERE", client);
 
 client.on("ready", async () => {
   dbl.serverCount();
   // console.log("Server count posted")
-  
   let hasVote = await dbl.hasVoted("714451348212678658");
   if(hasVote === true) {
-    console.log("Voted")
+    console.log("Voted");
   } else {
-    console.log("Vote please.")
+    console.log("Vote please.");
   }
   
   
-  let search = await dbl.search("779641401482805289")
-  console.log(search)
+  let search = await dbl.search("779641401482805289");
+  console.log(search);
   /*
   { RESULTS :
     avatar: 'https/cdn.discordapp.com/avatars/807855659173150781/f360c03f3ba6705c4d994b4edb19a634.webp?width=100&height=100',
